A function to display a list of colleagues who need to be congratulated on their birthday for the week.
Acceptance conditions
get_birthdays_per_week outputs birthdays in the format:
Monday: Bill, Jill
Friday: Kim, Jan

Users whose birthday was on the weekend should be congratulated on Monday.
For debugging, it is convenient to create a test list of users and fill it yourself.
The function displays users with birthdays one week ahead of the current day.
The week starts from Monday.